requirements.txt
tensorflow 
numpy 
pandas 
scikit-learn 
streamlit 
pillow 
opencv-python
